+++ test/dummy/app/campaigns/first_campaign.rb
@@ -1,4 +1,5 @@
 class FirstCampaign < Heya::Campaigns::Base
   step :one, subject: "First subject"
   step :two, subject: "Second subject"
+  step :three, subject: "Third subject"
 end