+++ test/dummy/db/migrate/20200214174703_create_heya_tables.rb
@@ -4,6 +4,7 @@ class CreateHeyaTables < ActiveRecord::Migration[6.0]
       t.references :user, null: false, polymorphic: true, index: false
 
       t.string :campaign_gid, null: false
+      t.string :step_gid, null: false
       t.boolean :concurrent, null: false, default: false
 
       t.datetime :last_sent_at, null: false